moved std offset and lib array to a shared folder
offset and array currently are free standing and thus can float between the kernel and userspace
