+++ src/kernel/boot/io.h
@@ -18,7 +18,7 @@
 #ifndef HEPHAIST_OS_DRIVERS_IO_IO_H
 #define HEPHAIST_OS_DRIVERS_IO_IO_H
 
-#include "kernel/lib/libc/stdoffset.h"
+#include <stdoffset.h>
 
 namespace kernel {
