+++ src/shared/lib/libc/stdoffset.h
