+++ src/shared/lib/libcpp/array.h
